requirements.txt
pandas==2.2.3
openpyxl==3.1.5
pypdf==5.4.0
openai==1.68.2
qdrant-client==1.17.0
python-dotenv==1.2.2
streamlit==1.44.1
